--- src/main/java/flab/transtalk/common/exception/NotFoundExceptionHandler.java
@@ -1,20 +1,0 @@
-package flab.transtalk.common.exception;
-
-import flab.transtalk.common.dto.res.ApiErrorResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@RestControllerAdvice
-public class NotFoundExceptionHandler {
-    @ExceptionHandler(NotFoundException.class)
-    public ResponseEntity<ApiErrorResponse> handleNotFoundException(NotFoundException e){
-        List<String> errors = new ArrayList<>();
-        errors.add(e.getMessage());
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ApiErrorResponse("NOT_FOUND", errors));
-    }
-}
