+++ src/main/java/flab/transtalk/common/exception/handler/CustomExceptionHandler.java
@@ -1,6 +1,8 @@
-package flab.transtalk.common.exception;
+package flab.transtalk.common.exception.handler;
 
 import flab.transtalk.common.dto.res.ApiErrorResponse;
+import flab.transtalk.common.exception.BadRequestException;
+import flab.transtalk.common.exception.NotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -10,8 +12,13 @@ import java.util.ArrayList;
 import java.util.List;
 
 @RestControllerAdvice
-public class BadRequestExceptionHandler {
-
+public class CustomExceptionHandler {
+    @ExceptionHandler(NotFoundException.class)
+    public ResponseEntity<ApiErrorResponse> handleNotFoundException(NotFoundException e){
+        List<String> errors = new ArrayList<>();
+        errors.add(e.getMessage());
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ApiErrorResponse("NOT_FOUND", errors));
+    }
     @ExceptionHandler(BadRequestException.class)
     public ResponseEntity<ApiErrorResponse> handlerBadRequestException(BadRequestException e){
         List<String> errors = new ArrayList<>();
